refactor: move application-info-form to new-application-dialog
