--- src/app/dashboard/applications/components/application-info-form/application-info-form.component.scss
@@ -1,3 +1,0 @@
-.margin-bottom {
-  margin-bottom: 16px;
-}
