--- src/app/dashboard/applications/components/application-info-form/application-info-form.module.ts
@@ -1,39 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { MatButtonModule } from '@angular/material/button';
-import { MatNativeDateModule } from '@angular/material/core';
-import { MatDatepickerModule } from '@angular/material/datepicker';
-import { MatDialogModule } from '@angular/material/dialog';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatIconModule } from '@angular/material/icon';
-import { MatInputModule } from '@angular/material/input';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { MatSelectModule } from '@angular/material/select';
-import { MatTooltipModule } from '@angular/material/tooltip';
-
-import { ApplicationInfoFormComponent } from './application-info-form.component';
-
-import { ConfirmationDialogModule } from '~components/confirmation-dialog/confirmation-dialog.module';
-
-@NgModule({
-  declarations: [ApplicationInfoFormComponent],
-  imports: [
-    CommonModule,
-    ConfirmationDialogModule,
-    FormsModule,
-    MatButtonModule,
-    MatDatepickerModule,
-    MatDialogModule,
-    MatFormFieldModule,
-    MatIconModule,
-    MatInputModule,
-    MatNativeDateModule,
-    MatProgressSpinnerModule,
-    MatSelectModule,
-    MatTooltipModule,
-    ReactiveFormsModule
-  ],
-  exports: [ApplicationInfoFormComponent]
-})
-export class ApplicationInfoFormModule {}
