+++ src/app/dashboard/applications/components/new-application-dialog/new-application-dialog.component.html
@@ -1,6 +1,101 @@
 <h1 mat-dialog-title class="at-text" [ngClass]="dialogData.column.color">New application</h1>
-<at-application-info-form
-  mat-dialog-content
-  [action]="dialogActions.New"
-  [currentColumn]="dialogData.column"
-></at-application-info-form>
+<mat-dialog-content>
+  <form [formGroup]="applicationForm">
+    <div class="form-grid">
+      <!-- Company Input -->
+      <mat-form-field class="full-width">
+        <mat-label>Company</mat-label>
+        <input formControlName="company" matInput type="text" />
+        <mat-error>{{ getError(company) }}</mat-error>
+      </mat-form-field>
+
+      <!-- Position Input -->
+      <mat-form-field class="full-width">
+        <mat-label>Position</mat-label>
+        <input formControlName="position" matInput type="text" />
+        <mat-error>{{ getError(position) }}</mat-error>
+      </mat-form-field>
+    </div>
+
+    <div [@expandCollapse]="viewMore ? 'expanded' : 'collapsed'">
+      <div @ngIfAnimation *ngIf="viewMore" class="form-grid">
+        <!-- Location Input -->
+        <mat-form-field class="full-width">
+          <mat-label>Location</mat-label>
+          <input formControlName="location" matInput type="text" />
+          <mat-error>{{ getError(location) }}</mat-error>
+        </mat-form-field>
+
+        <!-- Application Link Input -->
+        <mat-form-field class="full-width">
+          <mat-label>Application link</mat-label>
+          <input formControlName="link" matInput type="text" />
+          <mat-error>{{ getLinkError(link) }}</mat-error>
+        </mat-form-field>
+
+        <!-- Compensation Input -->
+        <div class="mat-form-field-combined">
+          <mat-form-field class="full-width" floatLabel="always">
+            <mat-label>Compensation</mat-label>
+            <input formControlName="compensation" matInput type="number" placeholder="0" class="number-input" />
+            <span matTextPrefix>$&nbsp;</span>
+            <span matTextSuffix>/</span>
+            <mat-error>{{ getError(compensation) }}</mat-error>
+          </mat-form-field>
+
+          <mat-form-field class="full-width">
+            <mat-label>Pay Period</mat-label>
+            <mat-select formControlName="payPeriod">
+              <mat-option *ngFor="let p of payPeriodOptions" [value]="p">
+                {{ p }}
+              </mat-option>
+            </mat-select>
+            <mat-error>{{ getError(payPeriod) }}</mat-error>
+            <mat-icon matSuffix>arrow_drop_down</mat-icon>
+          </mat-form-field>
+        </div>
+
+        <!-- Date Applied Date Picker -->
+        <mat-form-field class="full-width">
+          <mat-label>Date</mat-label>
+          <input formControlName="date" matInput [matDatepicker]="datePicker" [max]="currentDate" />
+          <mat-datepicker-toggle matSuffix [for]="datePicker">
+            <mat-icon matDatepickerToggleIcon class="at-text neutral"> today </mat-icon>
+          </mat-datepicker-toggle>
+          <mat-datepicker #datePicker></mat-datepicker>
+          <mat-error>{{ getError(date) }}</mat-error>
+        </mat-form-field>
+
+        <!-- Column Select -->
+        <mat-form-field class="full-width">
+          <mat-label>Column</mat-label>
+          <mat-select formControlName="column" [compareWith]="columnCompare">
+            <mat-option *ngFor="let col of columns$ | async" [value]="col" class="at-text" [ngClass]="col.color">
+              {{ col.title }}
+            </mat-option>
+          </mat-select>
+          <mat-error>{{ getError(column) }}</mat-error>
+          <mat-icon matSuffix>arrow_drop_down</mat-icon>
+        </mat-form-field>
+      </div>
+    </div>
+  </form>
+
+  <button *ngIf="!viewMore" mat-button (click)="toggleViewMore()" class="full-width margin-bottom">View more</button>
+
+  <!-- Buttons -->
+  <div class="button-row-container-end">
+    <button mat-button (click)="secondaryButtonClick()">Cancel</button>
+    <button
+      mat-flat-button
+      [color]="dialogData.column.color"
+      [disabled]="applicationForm.invalid || isLoading"
+      (click)="primaryButtonClick()"
+    >
+      <div *ngIf="isLoading; else text" class="loading-spinner-in-button">
+        <mat-spinner color="accent" diameter="24"></mat-spinner>
+      </div>
+      <ng-template #text>Save</ng-template>
+    </button>
+  </div>
+</mat-dialog-content>
