+++ src/app/dashboard/applications/components/new-application-dialog/new-application-dialog.component.scss
@@ -1,0 +1,3 @@
+:host .margin-bottom {
+  margin-bottom: 16px;
+}
