+++ src/app/dashboard/applications/components/new-application-dialog/new-application-dialog.module.ts
@@ -1,14 +1,39 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { MatButtonModule } from '@angular/material/button';
+import { MatNativeDateModule } from '@angular/material/core';
+import { MatDatepickerModule } from '@angular/material/datepicker';
 import { MatDialogModule } from '@angular/material/dialog';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatIconModule } from '@angular/material/icon';
+import { MatInputModule } from '@angular/material/input';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { MatSelectModule } from '@angular/material/select';
+import { MatTooltipModule } from '@angular/material/tooltip';
 
 import { NewApplicationDialogComponent } from './new-application-dialog.component';
 
-import { ApplicationInfoFormModule } from '~components/application-info-form/application-info-form.module';
+import { ConfirmationDialogModule } from '~components/confirmation-dialog/confirmation-dialog.module';
 
 @NgModule({
   declarations: [NewApplicationDialogComponent],
-  imports: [ApplicationInfoFormModule, CommonModule, MatDialogModule],
+  imports: [
+    CommonModule,
+    ConfirmationDialogModule,
+    FormsModule,
+    MatButtonModule,
+    MatDatepickerModule,
+    MatDialogModule,
+    MatFormFieldModule,
+    MatIconModule,
+    MatInputModule,
+    MatNativeDateModule,
+    MatProgressSpinnerModule,
+    MatSelectModule,
+    MatTooltipModule,
+    ReactiveFormsModule
+  ],
   exports: [NewApplicationDialogComponent]
 })
 export class NewApplicationDialogModule {}
